Refactored docs command code into a package
