+++ cmd/docs/doc.go
@@ -1,0 +1,2 @@
+// Package docs enjoys
+package docs
